% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc.table.R
\name{create.perc.table}
\alias{create.perc.table}
\title{Create percentile table function}
\usage{
create.perc.table(data)
}
\arguments{
\item{data}{Data.frame or matrix in which you want to know the percentiles for each column}
}
\description{
This function calculates percentiles by column for a given data set
}
\examples{
cat_function()
}
\keyword{perc}
