% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{estimate.ldl}
\alias{estimate.ldl}
\title{Predict LDL levels}
\usage{
estimate.ldl(data)
}
\arguments{
\item{data}{Data.frame or matrix where the first column is what you want to predict, and 
remaining columns are what you want to consider as possibly covariates. Any factors MUST be set as factors before 
running. No character vectors allowed.}
}
\value{
A vector of length equal to number of rows of input data with predicted values of LDL
}
\description{
This function uses LASSO to select predictors in order to run linear regression to predict LDL for individuals. 
Outputs a vector of predicted values of LDL corresponding on input
There are no checks for NAs or anythingelse and thus this isa development level function.
}
\examples{
response <- rnorm(1000)
goodpredictors <- 0.5*response + rnorm(1000,0.5)
badpredictors <- matrix(rnorm(10000),ncol=10)
estimate.ldl(cbind(response,goodpredictors,badpredictors))
}
\keyword{preds}
